Enable bf8 DPAS (#18)
Adds initial support for the native fp8 DPAS operation. Currently only
`bf8` (aka e5m2) is tested, `hf8` is waiting for IGC support.
* Introduces the `support_fp8_dpas` module attribute, currently the
attribute is set by looking for the FS1 device ID in the device name,
this is just a placeholder and will be replaced with an appropriate
device query.
* the upcast from `fp8` to `fp16` in `AccelerateMatmul` is now performed
conditionally based on the attribute above, this also required changes
in few other bits of code where we were assuming the upcast to take
place.