Set target_arch attribute for TTGIR tests. (#72)
All tests that provide TTGIR as an input for compilation fail for the
pISA target because the input module misses the target arch attribute,
and compilation always goes SPIRV path. We can modify tests to provide
proper target arch attributes or modify the compilation flow. I think it
would be more convenient for us to keep tests unchanged, so I propose to
simply set the target arch attribute in the `make_llir` phase of
compilation if this attribute doesn't exist.

Resolves #71

Signed-off-by: Ilya Enkovich <ilya.enkovich@intel.com>